+++ tripnest-backend/src/main/java/com/tripnest/tripnest_backend/repository/TripRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TripRepository extends JpaRepository<Trip, Integer> {
+public interface TripRepository extends JpaRepository<Trip, Long> {
+
     List<Trip> findByUserEmail(String email);
 }
